+++ cognite/neat/graph/transformers/_rdfpath.py
@@ -13,8 +13,8 @@ class ReduceHopTraversal(BaseTransformer):
     ...
 
 
-class AddAllReferences(BaseTransformer):
-    description: str = "Adds all references to specific property defined in Rules"
+class AddSelfReferenceProperty(BaseTransformer):
+    description: str = "Adds property that contains id of reference to all references of given class in Rules"
     _use_only_once: bool = True
     _need_changes = frozenset({})